fixed msvc warnings

+++ src/utils/bitstream.c
@@ -838,8 +838,8 @@ void gf_bs_get_content_no_truncate(GF_BitStream *bs, char **output, u32 *outSize
 		if (alloc_size) {
 			/*Align our buffer or we're dead!*/
 			gf_bs_align(bs);
-			*alloc_size = bs->size;
-			*outSize = bs->position;
+			*alloc_size = (u32) bs->size;
+			*outSize = (u32) bs->position;
 			*output = bs->original;
 		} else {
 			s32 copy = BS_CutBuffer(bs);
